[Feat]#2 loading창 navigationBar hidden
