+++ OpenMarket/OpenMarket/Prodict_List/ViewController/RegisterViewController.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 class RegisterViewController: UIViewController {
-
     override func viewDidLoad() {
         super.viewDidLoad()
         self.view.backgroundColor = .white
